Change Rust source files to includes
This change creates a `_includes` directory in the `_docs` directory and
adds includes for the relevant source files instead of copying them
directly into the markdown posts.

The `_includes` directory is actually a relative symlink to the patterns
directory, i.e,

    docs$ ls -l _includes
    total 4
    lrwxrwxrwx 1 nathan nathan 14 Jul 12 23:03 patterns -> ../../patterns

meaning it should work on any system that understands symlinks.

This means that changes to the Rust files are automatically reflected in
the rendered pages by Jekyll.

Closes #3.